requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
django-filter==22.1
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
pytils==0.4.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12